requirements.txt
# Universal Catalyst MCP Server Dependencies
fastmcp>=0.15.0
httpx>=0.27.0
python-dotenv>=1.0.0

# Knowledge pack system dependencies
jinja2>=3.1.6
pyyaml>=6.0.2
jq>=1.6.0

# Shared schema package (git submodule)
-e ./schemas

# Optional dependencies for different integration types
# Install only what you need for your specific knowledge packs

# Database integrations
asyncpg>=0.28.0        # PostgreSQL
aiomysql>=0.2.0        # MySQL
aiosqlite>=0.19.0      # SQLite  
motor>=3.3.0           # MongoDB
redis>=5.0.0           # Redis
elasticsearch>=8.0.0   # Elasticsearch

# Message queue integrations  
aio-pika>=9.3.0        # RabbitMQ
aiokafka>=0.8.0        # Apache Kafka
aioboto3>=12.0.0       # AWS SQS/S3

# File system integrations
# aioboto3>=12.0.0     # AWS S3 (already listed above)
google-cloud-storage>=2.10.0  # Google Cloud Storage
azure-storage-blob>=12.18.0   # Azure Blob Storage
aioftp>=0.21.0         # FTP
# asyncssh>=2.14.0     # SFTP (already listed below)

# SSH and shell integrations
asyncssh>=2.14.0       # SSH connections

# Python standard library packages used (no additional dependencies needed):
# - asyncio (built-in)
# - json (built-in) 
# - logging (built-in)
# - xml.etree.ElementTree (built-in)
# - dataclasses (built-in)
# - typing (built-in)
# - collections (built-in)
# - enum (built-in)
# - time (built-in)
# - uuid (built-in)
# - os (built-in)
# - sys (built-in)
# - signal (built-in)